moved all golang files to intermediate src directory
