+++ services/text-event-service/src/dataservice/filters.go
