+++ services/text-event-service/src/dataservice/memory.go
