+++ services/text-event-service/src/docs/errors.go
