+++ services/text-event-service/src/docs/responses.go
