+++ services/text-event-service/src/handlers/common_test.go
