+++ services/text-event-service/src/handlers/delete_test.go
