+++ services/text-event-service/src/handlers/docs.go
