+++ services/text-event-service/src/handlers/events_test.go
