+++ services/text-event-service/src/handlers/get_test.go
