+++ services/text-event-service/src/model/errors.go
