+++ services/text-event-service/src/model/errors_test.go
