+++ services/text-event-service/src/model/event_test.go
