+++ services/text-event-service/src/model/events_test.go
