+++ services/text-event-service/src/model/timespan.go
