+++ services/text-event-service/src/model/timespan_test.go
