+++ services/text-event-service/src/utils/validator_test.go
